Cambios a reservas y añadidas validaciones para reservas, clientes y usuarios

+++ src/main/java/com/example/Foodies/Reserva/dtos/ReservaListDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalTime;
 
 public record ReservaListDTO(
         Long id,
-        Long cantidad,
+        Integer cantidad,
         LocalTime horariollegada,
         EstadoReserva estadoReserva
 ) { }
